move unit tests to dedicated dir
