+++ domestic/management/commands/bootstrap_great.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.core.management import BaseCommand
 from wagtail.core.models import Page, Site
 
-from tests.domestic import factories
+from tests.unit.domestic import factories
 
 User = get_user_model()
