+++ tests/unit/core/test_views.py
